refactor: extract config JSON pointer helpers
