+++ src/core/config/json_pointer.rs
@@ -1,0 +1,257 @@
+use crate::core::error::Error;
+use crate::core::Result;
+use serde_json::Value;
+
+pub fn set_json_pointer(root: &mut Value, pointer: &str, new_value: Value) -> Result<()> {
+    let pointer = normalize_pointer(pointer)?;
+    let Some((parent_ptr, token)) = split_parent_pointer(&pointer) else {
+        *root = new_value;
+        return Ok(());
+    };
+
+    let parent = ensure_pointer_container(root, &parent_ptr)?;
+    set_child(parent, &token, new_value)
+}
+
+/// Remove the value at a JSON pointer path.
+pub fn remove_json_pointer(root: &mut Value, pointer: &str) -> Result<()> {
+    let pointer = normalize_pointer(pointer)?;
+    let Some((parent_ptr, token)) = split_parent_pointer(&pointer) else {
+        return Err(Error::validation_invalid_argument(
+            "pointer",
+            "Cannot remove root element",
+            None,
+            None,
+        ));
+    };
+
+    let parent = navigate_pointer(root, &parent_ptr)?;
+    remove_child(parent, &token)
+}
+
+/// Navigate to the value at a JSON pointer without creating intermediate objects.
+/// Returns an error if any segment along the path is missing.
+fn navigate_pointer<'a>(root: &'a mut Value, pointer: &str) -> Result<&'a mut Value> {
+    if pointer.is_empty() {
+        return Ok(root);
+    }
+
+    let tokens: Vec<String> = pointer.split('/').skip(1).map(unescape_token).collect();
+    let mut current = root;
+
+    for token in &tokens {
+        current = match current {
+            Value::Object(map) => map.get_mut(token.as_str()).ok_or_else(|| {
+                Error::validation_invalid_argument(
+                    "pointer",
+                    format!("Key '{}' not found", token),
+                    None,
+                    None,
+                )
+            })?,
+            Value::Array(arr) => {
+                let index = parse_array_index(token)?;
+                let len = arr.len();
+                if index >= len {
+                    return Err(Error::validation_invalid_argument(
+                        "pointer",
+                        format!("Array index {} out of bounds (length {})", index, len),
+                        None,
+                        None,
+                    ));
+                }
+                &mut arr[index]
+            }
+            _ => {
+                return Err(Error::validation_invalid_argument(
+                    "pointer",
+                    format!("Cannot navigate through non-object at path: {}", pointer),
+                    None,
+                    None,
+                ))
+            }
+        };
+    }
+
+    Ok(current)
+}
+
+fn remove_child(parent: &mut Value, token: &str) -> Result<()> {
+    match parent {
+        Value::Object(map) => {
+            if map.remove(token).is_none() {
+                return Err(Error::validation_invalid_argument(
+                    "pointer",
+                    format!("Key '{}' not found", token),
+                    None,
+                    None,
+                ));
+            }
+            Ok(())
+        }
+        Value::Array(arr) => {
+            let index = parse_array_index(token)?;
+            if index >= arr.len() {
+                return Err(Error::validation_invalid_argument(
+                    "pointer",
+                    format!("Array index {} out of bounds (length {})", index, arr.len()),
+                    None,
+                    None,
+                ));
+            }
+            arr.remove(index);
+            Ok(())
+        }
+        _ => Err(Error::validation_invalid_argument(
+            "pointer",
+            "Cannot remove from non-container type",
+            None,
+            None,
+        )),
+    }
+}
+
+fn normalize_pointer(pointer: &str) -> Result<String> {
+    if pointer.is_empty() {
+        return Ok(String::new());
+    }
+
+    if pointer == "/" {
+        return Err(Error::validation_invalid_argument(
+            "pointer",
+            "Invalid JSON pointer '/'",
+            None,
+            None,
+        ));
+    }
+
+    if !pointer.starts_with('/') {
+        return Err(Error::validation_invalid_argument(
+            "pointer",
+            format!("JSON pointer must start with '/': {}", pointer),
+            None,
+            None,
+        ));
+    }
+
+    Ok(pointer.to_string())
+}
+
+fn split_parent_pointer(pointer: &str) -> Option<(String, String)> {
+    if pointer.is_empty() {
+        return None;
+    }
+
+    let mut parts = pointer.rsplitn(2, '/');
+    let token = parts.next()?.to_string();
+    let parent = parts.next().unwrap_or("");
+
+    let parent_ptr = if parent.is_empty() {
+        String::new()
+    } else {
+        parent.to_string()
+    };
+
+    Some((parent_ptr, unescape_token(&token)))
+}
+
+fn ensure_pointer_container<'a>(root: &'a mut Value, pointer: &str) -> Result<&'a mut Value> {
+    if pointer.is_empty() {
+        return Ok(root);
+    }
+
+    let tokens: Vec<String> = pointer.split('/').skip(1).map(unescape_token).collect();
+
+    let mut current = root;
+
+    for token in tokens {
+        let next = match current {
+            Value::Object(map) => map
+                .entry(token)
+                .or_insert_with(|| Value::Object(serde_json::Map::new())),
+            Value::Null => {
+                *current = Value::Object(serde_json::Map::new());
+                if let Value::Object(map) = current {
+                    map.entry(token)
+                        .or_insert_with(|| Value::Object(serde_json::Map::new()))
+                } else {
+                    unreachable!()
+                }
+            }
+            Value::Array(arr) => {
+                let index = parse_array_index(&token)?;
+                if index >= arr.len() {
+                    return Err(Error::config_invalid_value(
+                        pointer,
+                        None,
+                        "Array index out of bounds while creating path",
+                    ));
+                }
+                &mut arr[index]
+            }
+            _ => {
+                return Err(Error::config_invalid_value(
+                    pointer,
+                    Some(value_type_name(current).to_string()),
+                    "Expected object/array at pointer",
+                ))
+            }
+        };
+
+        current = next;
+    }
+
+    Ok(current)
+}
+
+fn set_child(parent: &mut Value, token: &str, value: Value) -> Result<()> {
+    match parent {
+        Value::Object(map) => {
+            map.insert(token.to_string(), value);
+            Ok(())
+        }
+        Value::Array(arr) => {
+            let index = parse_array_index(token)?;
+            if index >= arr.len() {
+                return Err(Error::config_invalid_value(
+                    "arrayIndex",
+                    Some(index.to_string()),
+                    "Array index out of bounds",
+                ));
+            }
+            arr[index] = value;
+            Ok(())
+        }
+        _ => Err(Error::config_invalid_value(
+            "jsonPointer",
+            Some(value_type_name(parent).to_string()),
+            "Cannot set child on non-container",
+        )),
+    }
+}
+
+fn parse_array_index(token: &str) -> Result<usize> {
+    token.parse::<usize>().map_err(|_| {
+        Error::validation_invalid_argument(
+            "arrayIndex",
+            "Invalid array index token",
+            Some(token.to_string()),
+            None,
+        )
+    })
+}
+
+fn unescape_token(token: &str) -> String {
+    token.replace("~1", "/").replace("~0", "~")
+}
+
+fn value_type_name(value: &Value) -> &'static str {
+    match value {
+        Value::Null => "null",
+        Value::Bool(_) => "bool",
+        Value::Number(_) => "number",
+        Value::String(_) => "string",
+        Value::Array(_) => "array",
+        Value::Object(_) => "object",
+    }
+}
